add simple markdown editor
this change adds js script and styles for the simple markdown editor,
which is replacing the existing editable paragraph element. the
simpleMDE is instantiated and passed to the Detail view where it is
used to set and retrieve the value from the editor using simpleMDE.value

the event handling in the Detail element had to be changed because
simpleMDE does not have an easy way to add a data-name attribute to it.
Therefore, event listeners are added directly to each of the existing
data-name elements (title, published), while the event listener for
the body is added with simpleMDE.codemirror.on('change', ...)